writes diagnostics file in user's home directory
